requirements.txt
ansi2html==1.8.0
appnope==0.1.3
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
dash==2.8.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.6
decorator==5.1.1
executing==1.2.0
fastjsonschema==2.16.2
findspark==2.0.1
Flask==2.2.3
fonttools==4.38.0
idna==3.4
importlib-metadata==6.0.0
ipykernel==6.21.2
ipython==8.10.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter-dash==0.4.2
jupyter_client==8.0.2
jupyter_core==5.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
mysql-connector-python==8.0.32
nbformat==5.7.3
nest-asyncio==1.5.6
numpy==1.24.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
pathspec==0.11.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
plotly==5.13.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.5
pyarrow==11.0.0
Pygments==2.14.0
PyInputPlus==0.2.12
pyparsing==3.0.9
pyrsistent==0.19.3
PySimpleValidate==0.2.12
pyspark==3.3.1
python-dateutil==2.8.2
pytz==2022.7.1
pyzmq==25.0.0
requests==2.28.2
retrying==1.3.4
scikit-learn==1.2.1
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
stack-data==0.6.2
stdiomask==0.0.6
tenacity==8.2.1
threadpoolctl==3.1.0
tokenize-rt==5.0.0
tomli==2.0.1
tornado==6.2
traitlets==5.9.0
typing_extensions==4.4.0
urllib3==1.26.14
wcwidth==0.2.6
Werkzeug==2.2.3
zipp==3.13.0
